requirements.txt
argparse
dill
fire
git+https://github.com/ppwwyyxx/tensorpack.git
Flask==0.10.1
itsdangerous==0.24
paho-mqtt==1.4.0
absl-py==0.5.0
astor==0.7.1
backcall==0.1.0
bleach==2.1.4
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
colorama==0.3.9
cycler==0.10.0
Cython==0.28.5
dask==0.19.2
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.1
django-heroku==0.3.1
entrypoints==0.2.3
ffmpeg==1.2.1
gast==0.2.0
grpcio==1.15.0
gunicorn==19.9.0
h5py==2.8.0
heroku==0.1.4
html5lib==1.0.1
idna==2.7
imageio==2.4.1
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0
MarkupSafe==1.0
matlab==0.1
matplotlib==3.0.0
mistune==0.8.3
msgpack==0.5.6
msgpack-numpy==0.4.3.2
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numpy==1.14.5
opencv-python==3.4.3.18
paho-mqtt==1.4.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
pipenv==2018.7.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.7
psycopg2==2.7.5
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.1
PyQt5==5.11.2
PyQt5-sip==4.19.12
pyqtgraph==0.10.0
pyreadline==2.1
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
pywinpty==0.5.4
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
slidingwindow==0.0.13
tabulate==0.8.2
tensorboard==1.10.0
tensorflow==1.10.0
tensorflow-gpu==1.10.0
tensorpack==0.8.9
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wget==3.2
whitenoise==4.1
widgetsnbextension==3.4.2
